pyproject.toml
[tool.poetry]
name = "freton-backend"
version = "0.1.0"
description = ""
authors = ["Antonio Fernandes <as.fernandes.jr@gmail.com>"]
readme = "README.md"
packages = [{ include = "freton" }]

[tool.poetry.dependencies]
python = "^3.10"
markdown = "^3.5.2"
django-filter = "^24.1"
djangorestframework = "^3.14.0"
django = "^4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
